Organize code: clean up aiChat tool, move featureFlags, formatResponse, and typeGuards helpers to utils folder
