+++ src/tools/index.ts
@@ -1,3 +1,4 @@
+export { registerAiChatTools } from "./aiChat";
 export { registerCmsTools } from "./cms";
 export { registerComponentsTools } from "./components";
 export { registerPagesTools } from "./pages";
