+++ src/utils/formatResponse.ts
@@ -20,9 +20,3 @@ export function formatErrorResponse(error: any) {
     ],
   };
 }
-
-export function isApiError(
-  error: unknown
-): error is { status: number; message?: string } {
-  return typeof error === "object" && error !== null && "status" in error;
-}
