+++ src/utils/typeGuards.ts
@@ -1,0 +1,8 @@
+interface ApiError {
+  status: number;
+  message?: string;
+}
+
+export function isApiError(error: unknown): error is ApiError {
+  return typeof error === "object" && error !== null && "status" in error;
+}
